Remove simd_json benchmarks
The benchmarks don't show much improvement over standard serde.

This is probably due to the fact that most objects are not very large and also because we are not using the `simd_json_derive` impls.

The latter are difficult to implement without making the already complex attribute configurations even more convoluted. This is because `simd_json_derive` doesn't support several of the `serde` and `serde-as` features we are using (e.g., `rename = "kebab-case"`, flatten, etc).

Since it's unlikely that `simd_json` will make much of a difference anyway, it doesn't seem worth adding special support for. But maybe revisit later.